+++ viadot/tasks/__init__.py
@@ -47,5 +47,6 @@ from .epicor import EpicorOrdersToDF
 
 from .sftp import SftpToDF, SftpList
 from .mindful import MindfulToCSV
+from .hubspot import HubspotToDF
 from .mediatool import MediatoolToDF
 from .customer_gauge import CustomerGaugeToDF