+++ CHANGELOG.md
@@ -6,10 +6,11 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 ### Added
+- Fixed `_full_trfresh` method in `Databricks` source.
 - Enabled Databricks Connect in the image. To enable, [follow this guide](./README.md#executing-spark-jobs)
 - Added `Databricks` source to the library.
 - Added `ExchangeRates` source to the library.
-- Added `from_df()` method to `Azure Data Lake` source
+- Added `from_df()` method to `Azure Data Lake` source.
 - Added `SAPRFC` source to the library.
 - Added `S3` source to the library.
 - Added `RedshiftSpectrum` source to the library.