+++ CHANGELOG.md
@@ -6,6 +6,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 - Added `anonymize_df` task function to `task_utils.py` to anonymize data in the dataframe in selected columns.
+- Added `CustomerGauge` source class
+- Added `CustomerGaugeToDF` task class
 
 ## [0.4.13] - 2023-03-15
 ### Added