+++ CHANGELOG.md
@@ -9,6 +9,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Added `anonymize_df` task function to `task_utils.py` to anonymize data in the dataframe in selected columns.
 - Added `CustomerGauge` source class
 - Added `CustomerGaugeToDF` task class
+- Added `CustomerGaugeToADLS` flow class
 
 ## [0.4.13] - 2023-03-15
 ### Added